requirements.txt
openai>=1.40.0
anthropic>=0.39.0
PyYAML>=6.0.0
requests>=2.31.0
python-dotenv>=1.0.1